feat: level component
