+++ src/components/level/index.ts
@@ -1,4 +1,5 @@
 export * from "./agent";
+export * from "./level";
 export * from "./pathfinder";
 export * from "./patrol";
 export * from "./sentry";
