+++ src/index.ts
@@ -59,6 +59,8 @@ export type {
     ForceMode,
     HealthComp,
     LayerComp,
+    LevelComp,
+    LevelOpt,
     LifespanCompOpt,
     MaskComp,
     NamedComp,
@@ -72,6 +74,7 @@ export type {
     PathfinderCompOpt,
     PathfinderMapComp,
     PathfinderMapCompOpt,
+    PathFindOpt,
     PatrolComp,
     PatrolCompOpt,
     PlatformEffectorComp,
@@ -111,7 +114,6 @@ export type {
     GameObjEventMap,
     GameObjEventNames,
     KeepFlags,
-    LevelOpt,
     SceneDef,
     SceneName,
     SetParentOpt,
@@ -203,7 +205,6 @@ export type {
     KGamepad,
     KGamepadButton,
     LerpValue,
-    LevelComp,
     LoadFontOpt,
     Mask,
     MergeComps,
@@ -212,7 +213,6 @@ export type {
     MouseButton,
     MusicData,
     Outline,
-    PathFindOpt,
     PluginList,
     QueryOpt,
     Recording,
