+++ src/kaplay.ts
@@ -204,6 +204,7 @@ import {
     follow,
     health,
     layer,
+    level,
     lifespan,
     mask,
     move,
@@ -1362,6 +1363,7 @@ const kaplay = <
         sentry,
         patrol,
         pathfinder,
+        level,
         fakeMouse,
         // group events
         trigger,
